sixth commit

+++ src/toko/Handphone.java
@@ -1,13 +1,15 @@
 package toko;
 
-public class Handphone extends Barang{
+public class Handphone extends Barang {
     private String warna;
     public static int total = 0;
 
-    public Handphone(int id, String nama, int harga, int stok, String warna){
+    public Handphone(int id, String nama, double harga, int stok, String warna) {
         super(id, nama, harga, stok);
         this.warna = warna;
     }
 
-    public String getWarna() {return warna;}
+    public String getWarna() {
+        return warna;
+    }
 }
